Stream repositories to and from api and cache
Prior to this change, we would request _all_ the repositories for the
current user before we displayed them. For most users, this won't be a
problem.

For users like me, it's a big problem. The dialog appears to hang. With
this change, we take the same approach as GitHub for Windows and we
request repositories in chunks.

- Current User's owned repos
- Current User's member repos
- Repos for each org the user is a member of

Note that each individual grouping could represent multiple HTTP request
because of paging. We still aggregate all the requests for each grouping
and cache that all under a single cache key. Even so, this greatly
increases the responsiveness of the Clone dialog.