Fix NullRef when viewingcontrols in the designer
Looks like Xaml was having trouble with the ArrayExtension. Instead, I
used the xaml to bind to the design time DataContext. Since we also bind
in the code-behind, I believe at runtime the codebehind takes
precedence, though in truth it probably doesn't matter because we set
the DataContext correctly in both cases.

Fixes #163