Ensure we dispose things properly
When we login (from cache or otherwise) we add the RepositoryHost to the
disposables collection to be disposed when the RepositoryHost is
disposed. When we logout, we dispose of the host by removing it from the
disposables collection.

When `PackageBase` is disposed, we dispose `RepositoryHost` which
handles just about everything.

The tricky part is figuring out who should dispose what. For example,
multiple classes import `ISharedCache`, but I had

Fixes #153