Make the dispose pattern the same across all our classes, for consistency.
We currently don't have native resources to dispose so we don't have any
finalizers that call Dispose(false), making most of these checks
redundant, but it's easier to just have the pattern implemented in the
correct safe way so it never fails and is easy to copy paste around.

I debated whether to have the if(disposed) check inside or outside the
if (disposing) check. Since we're only disposing managed resources, and
some of the dispose code will only run safely if *not* called from a finalizer
(it's not safe if disposing == false), I'm treating disposed as a
flag for having managed resources disposed. If we ever do need finalizers,
we can revisit this on a case-by-case basis.