Fix race condition loading the host list, and also resetting notifications
There's a bit of a race condition when loading the control for the first
time, the hosts are in the process of logging in and might not have the
IsLoggedIn flag set yet. We can trust that all the connections in the
ConnectionManager are valid and are the ones we'll use, so use that list
to populate the hosts dropdown. By the time we need the actual host,
it'll be ready.

Also, reset notifications that are host-specific when changing host.