Ensure we only return one user
We had one case where the user's account was listed twice. I was unable
to repro it and as far as I can tell, it should be impossible.

I found one potential case where it could happen, but it's a case that
should, as far as I can tell, never happen. But every time I say that,
it happens. So I put in a defensive check.
Hopefully Fixes #334