Fallback if initializing BlobCache fails
We hope this never happens, but Akavache uses Sqlite and you never know
if some roque process or thread interferes and causes it to be bad.

So when initializing the SharedCache, if anything fails we replace it
with the InMemoryBlobCache. This is fine for our cases because the only
cache that persists information that we can't recreate is the
SecureCache. And hopefully that never fails because we're using the
Windows Credential Storage rather than Akavache for that.

This sort of Fixes #357, though it's more of a band-aid.