Fix loading resources assemblies
There are some views that load very early, before VS has had a chance to
set the proper load contexts. In some situations, it fails to load
GitHub.UI when referenced from xaml because of this.
SharedDictionaryManager has an assembly resolver, but it was in
GitHub.UI so that wasn't going to help much.

Move SharedDictionaryManager to the main assembly and merge it with the
CachedResourceDictionary so that anything that tries to load resources
from our assemblies is guaranteed to go through our assembly resolver,
which knows where things are (in case VS is confused and doesn't know
where things are).

Fixes #382