Fix slowdowns when showing nav items
Should finally fix #397

Instead of having the nav items receive notification about which repo is
loaded (so they can properly update themselves), have the home section
(which loads first) propagate that information to the nav items (because
it has it).

This seems to fix the slowdowns, perhaps because now only one class is
listening to the PropertyChanged event that is triggered when the active
repo changes, instead of all of them.

Also, disable the issues button for now, until I figure out if the delay-load
is affecting things.