Don't save the repo list in the cache
The repository list is loaded from the registry into ConnectionManager.
Don't save it in the cache for now, so it doesn't break compatibility
with the existing cache files and we don't really need to cache it right
now. Let's see how it works just loading directly from the registry for
now.