Fix reacting to new repos on the list
If the user is logged out and clicks Create or Clone, they'll go through
the login dialog and the repository list will get populated as soon as
that's done, before the actual create/clone action starts. We don't want
to handle those "new" repos on the list as if they've just been
cloned/created.