Show custom notifications for cloned repos too
Now both creating and cloning show messages with the url of the
repository, plus a link to create or open a project/solution, depending
on the situation.
If it's a created repo, link to the Create dialog.
If it's a cloned repo that doesn't have commits, link to the Create
dialog.
If it's a cloned repo that has commits, link to the Open dialog.