Add more functionality to SimpleRepositoryModel
Now that ActiveRepo is a SimpleRepositoryModel, we need a way to update
the url, for when a user adds or removes the origin remote manually
inside Visual Studio (in the git repo settings), or via the command
line.
This is required so that the UI updates properly and rechecks whether
the repo is ours or not.