Add support for testing code that uses LibGit2Sharp
Testing RepositoryModel creation via the path constructor implies
hitting libgit2sharp codepaths, and libgit2sharp doesn't expose nice
interfaces for testing (why??). So, move things around and add a
GitService that replaces the extension methods that depend on
libgit2sharp, so we can at least mock at a higher level.

The things that require this GitService are not created via mef, so
expose it via the Services static class.