Stream results from PR api as they come in
We now fetch initial results from cache, streaming them into an
observable collection. Then we fetch updated data online, streaming
those into the same collection so that the items get updated as they
come in. The UI shows data as it comes in and refreshes as live data is
updated.