A cache index for akavache objects
CacheIndex stores a list of string keys of all the objects that are a
part of a query - all pull requests for a repository (regardless of
state); all repositories for a user, etc. When a query is performed
(GetPullRequests for instance), the CacheIndex object is obtained from
akavache and all the objects referenced by the index keys are returned.
If it needs to be refreshed, the fetch function is called and each
returned object will be added/refreshed in akavache and the index will
also be updated.

This way, no matter what type of filtering is performed on the query,
only one object of each type will be saved (keys are in the form
`user|repo|objecttype|id`), and we can leverage the fact that GitHub
list apis return full objects and not trimmed-down indexes, avoiding
doing the same query twice (once for list, once for detail), since the
object will be fully cached during listing and immediately available for
detail.

We will want to use sql queries directly at some point, for searches and
other more advanced uses, but for now this works. We can use these index
objects to create real sql indexes later on.