Do not mix reference and value checks
In equality methods and operators, I prefer not to mix reference and
value checks in one line. These are different checks and the value parts
may change in the future. These are very easy to get wrong, so I prefer
to keep them as readable and split up as possible.