Allow replacing the IConnection service registration
For concurrent ui flows, we register the connection of the current flow
which is immediately picked up by the viewmodel. Another ui flow might
run and register a connection to be picked up by its viewmodel, and by
the time the first ui flow is done and tries to remove the iconnection
registration, that iconnection is not the one it registered any more. So
we're not removing registrations any more, and we just let the
UIController replace the IConnection any time it needs to launch a new
viewmodel.