Fix showing auth dialogs for long-running ui flows
We already support showing a separate dialog for authentication, but it
only worked when the selected ui flow was either authentication or the
clone/create flows, which reuse the same dialog window. It also worked
for publish, but that one is actually cheating by running two separate
ui flows (the auth and then the publish)

Since we now support long running ui flows that switch to subflows when
needed, we need to be able to show a view in either the pane or in the
dialog, basically popping up the dialog whenever needed and then
shutting it down without stopping the ui flow.