Extend canCreateGist checking to also look at the selected text.
I was not restricting this before, but if we try to send a new gist with
the file content of it as null, an empty string or white space octokit
returns a validation error saying the files field is missing.  This is
fine, but the Create Gist UI needs to be updated to show that no text
was selected.