Make sure an IConnection is always registered
Fix a bug where if we're logged in but don't pass a connection to Start,
no connection object would be passed along to whatever viewmodel we're
loading, which is a Bad Thing(tm).

If we're logged in but no connection object is passed in, we should
retrieve the first available logged-in connection from the connections
list and pass that one in, so that viewmodels *always* get a connection.