ViewModels only notify if there are errors
If operations are successful, we know from the result of the command, so
ViewModels only call the notification service in case of errors (so we
can grab the message). Also, they shouldn't be hooking directly into the
VS UI, that's for the view to handle (or whoever is controlling the
view, if the view is embedded and the error needs to be shown somewhere
else)

Fix tests so that they look for the correct events when things change
and not rely on the notification service for anything other than errors.