Listen to all relevant notifications
There's two types of notifications that can come up via the Publish
viewmodel: warnings from validators while filling out the form, and
errors from the publish process. Errors are sent via the notification
system, and warnings come up via the validators.

We're doing it like this for now because warnings are reactive based on
user input (so they go away if the validator becomes valid), and I'm
still not sure if the NotificationDispatcher should have a
`ClearNotifications` method (it's just dispatching messages, is it its
job to invalidate messages too? unsure)