Make sure only one instance of a service is created
IVSServices and ITeamExplorerServices have methods that depend on types
from specific versions of TeamFoundation. Unfortunately, because the
class itself doesn't depend on any TeamFoundation types, it can be
happily instantiated by the runtime, which means that there would be two
instances of each of these MEF exports floating around, which of course
silently crashes the MEF loading of any other types that depend on these
services in their constructors.

We need to make sure that any service exported by these
TeamFoundation-specific projects have fields/properties/constructors
that depend on TeamFoundation types, so that they will not be
instantiated if the VS they're running on is a different version.