Add a xaml behavior and converter to support [None] items on top of lists
For lists where we want to support having no selection, it's handy to
have an "empty" item on top of the list for the user to select. The
behavior added here will add such an item to the top of a list if the
selected property has a value set in it. If the selected property is
null, the behavior removes the "empty" item.

The behavior requires that the `ItemsSource` binding of the ItemsControl
it's attached to be a MultiBinding, where the first entry is the bound
property that contains the "empty" item to show.

The `StickieListItemConverter` makes sure that the ItemsSource gets set
with the second entry of the MultiBinding.