Fix all the bindings
A bunch of classes were doing xaml bindings in a very wrong way, relying
on some external code setting the ViewModel property explicitly and not
defining the bindings in xaml.
Make every SimpleUserControl subclass set the ViewModel property via the
DataContextChanged event (DataContext is set for every view in
UIController), and fix the xaml bindings so things work.
Remove useless ReactiveUI code-behind bindings while we're at it, it's
pointless doing it twice.

This should fix #274 and #275