Fix premature cache disposal
Now that we're not creating a million copies of the same cache every
time something wants something for the cache, `ModelService` can't just
willy-nilly dispose of the `hostCache` object when it feels like it,
because that cache object is in itself cached in the factory and will be
reused by other services, if needed.

The factory is the one responsible for keeping the lifetime of its
cached objects (and yes, now it should probably not be called a factory
anymore, it should be a service but oh well), and so now it needs to be
disposable and take care of things.