Make sure github pane is reentrant
Fixes #282

We were getting into a state when loading the pane for the first time
where `Reload` would be called once as a result of the `RepoChanged`
call, which triggers when `Initialize(ServiceProvider)` runs, and then
`Reload` would run again as a result of the `Initialize(ViewWithData)`
call, because the Pull Requests button was clicked, which causes the
pane to reset itself. These two `Reload` calls would be executing
concurrently, causing a race condition that would crash randomly if the
controller was stopped and restarted at just the right point.

Need to make sure that nothing else can happen while
`Initialize(ServiceProvider)` is running, and that
`Initialize(ServiceProvider)` is run serially (no asyncs here!).