Keep validation adornment visibility up to date
The `ShowValidateError` method wasn't getting called to update the
validation adorner when `ValidationMessage.ShowError` changes
independent of a `TextChanged` or focus event.

I'm not sure if this should be done here or in the `WhenAny` above which
updates `ShowError` - I've put it here as that's where the existing call
to ShowValidateError is.