Initial support for collapsing in clone dialog.
Makes groups in the clone dialog collapsible. If there is only a single
group, then it will be expanded; as more groups are created, the first
group is collapsed and further groups will be added collapsed.

Only problem with this is that you can see an initial flash of expanded
items when loading with multiple groups, however I felt this was
preferable to everything being collapsed until the whole list is
loaded...