Use Merge instead of Amb on login control.
Fixes #309. `Observable.Amb` "emits all of the items from only the first
of these Observables to emit an item or notification". This means that
if the user tries to login first to .com with an invalid login, then
tries to login to enterprise with a valid login the login from
enterprise will be ignored because the .com observable fired first.
Instead use `Observable.Merge` which will emit authentication results
from both in any order.