Removed DefaultRepositoryName
From `IRepositoryPublishViewModel`. Seems to not be needed and was
causing #279.

The default repository name was being placed directly into
`RepositoryPublishControl.nameText` in the control which meant:

- it was only written to `RepositoryName` when the user clicked on the
`nameText` control
- which meant that `RepositoryName` was up until that point null
- which meant that the validation for it was failing
- which meant the "Publish" button was disabled.

This commit just sets the `RepositoryName` to the default in the VM ctor
meaning that that name gets immediately validated.
`DefaultRepositoryName` didn't seem to be used anywhere else, so I
removed it.

cc: @Haacked for a sanity check because it looks like you wrote this.

Fixes #279