Make IO operations go through Funcs
This is the same pattern used by `ConnectionManager` which allows us to
pass in mocked rothko interfaces for testing, but ensure that the rothko
dependency isn't loaded at runtime. However in GitHub.Exports we don't
have and (possibly) don't want a dependency on rothko so we need to work
out what to do there.