Fix LinkDropDown.
To allow it to work with data templates. This removes the previous hack
in `UpdateLinkText` and adds a `LinkItemTemplate` property that can be
used to supply a data template with which to display the selected item.

Also had to modify `BranchNameConverter` slightly to accept plain
strings as input as when the `LinkDropDown` is using the `Header`
property as its `LinkItem` this will get passed through the converter
too.