Handle diff/open of non-checked out files.
When viewing a PR from a fork in the PR details view, if the fork has
not been fetched locally then "Compare File" and "Open File" were
failing silently. This is because to get the files to compare, we looked
up the PR HEAD commit and get the file from the returned tree. We did a
fetch before looking up the PR HEAD but if the fork is a repository that
does not yet have a remote then we won't have fetched the commit.

This adds a fallback to using the API and storing a cached version of
the file contents.