Correctly refresh repositories.
The previous logic was wrong in that it wouldn't refresh the enterprise
account repository list. Attempt to fix that here.

I've tried to only call `connectionManager.RefreshRepositories` once by
checking if `sectionIndex == 0` but I'm not partiularly happy with this
- it seems like the wrong place to be refreshing the connection manager
repositories, and if another section of a type other than a connect
section were to be added at index 0 then it would break.