Refactor Button styling.
- Remove `GitHubButton` style - this was the default style for all our
buttons, but wasn't used anywhere. It appears to be the default style in
GHfW but doesn't fit with Visual Studio. By removing this default style,
buttons in Team Explorer use the default button style - fixes #796.
- Make `GitHubVsButton` the default style for buttons in the GitHub
pane. We only currently have one button that's not a primary button
("Cancel" in the PR creation view) but it should be the default style
for subsequent buttons there, so removed the style from that individual
button and made it apply to all buttons.