Added IHasDone and IHasCancel interfaces.
Refactor the `Done` and `Cancel` signals from `IDialogView` into
`IHasDone` and `IHasCancel` interfaces so that views can define their
capabilities in a more granular fashion. This has the slightly
unfortunate side-effect of leaving `IDialogView` with only a single
`IsBusy` signal which can't be refactored out into an interface because
we already have an incompatible `IHasBusy` which is implemented by the
view models. This will be resolved in the next step of refactoring where
the signals are moved from the views to the view models.