Merge SimpleUserControl and BusyStateView.
Merges `SimpleUserControl` and `BusyStateView` into a single `ViewBase`
control. This control will automatically display a spinner if the view
model implements `IHasLoading` and `IsLoading` is true, and an
intermediate progress bar if the view model implements `IHasBusy` and
`IsBusy` is true. If the view model implements neither of these
interfaces, then it will act the same as a standard `UserControl`.