Markdown should not be included in the shared dictionary
The SharedDictionary in GitHub.UI is the lowest level of controls and is included in all the Team Explorer views (Home, Connect, etc). Therefore it should not bring in any external dependencies that cause other libraries to get loaded. Since markdown.xaml requires Markdig now, it should be referenced only in the views that use it.