Add message for pending review comments.
If the user has pending review comments on the server then we can't post new comments. The correct way to test for this would be to make a request to `/repos/:owner/:repo/pulls/:number/reviews` and check for comments with a `PENDING` state. For the moment however we'll just display a message.