No TargetVersion for GitHub.TeamFoundation.15
Don't specify a target Visual Studio version for the
GitHub.TeamFoundation.15 MEF assembly.

Specifying a TargetVersion was causing it to sometimes incorrectly be
excluded when installed on some Visual Studio 2017 versions.

We can leave TargetVersion on the GitHub.TeamFoundation.14 MEF assembly
because this attribute isn't supported on Visual Studio 2015 and we do
want it excluded on Visual Studio 2017.

This is a workaround for #1859.