Consolidate PathError into PathWarning
There was an issue with PathError appearing at the top. When the error
appeared, the repository list would move down causing a different
repository to be selected. This would clause flickering between two
repositories.

This commit consolidates PathError into PathWarning (which appears
above the path which is being validated).  This avoids the flicker and
puts the warning next to the text box which it applies to.

The logic to block clone or open when there is a file in the way has
been moved to the commands' `canExecute` observable.