Optimize when forked from local branch
Check for the common case where a branch was forked from a local
branch. Use CommonAncestor because we don't want to search for a commit
that only exists locally or that has been added to the remote tracking
branch since the fork. This won't work if new commits have been pulled
to forked from branch.