Fix calling IGitActionsExt.CloneAsync from VS 2019
The ServiceProgressData type is in a Visual Studio 2019 assembly that
we don't currently have access to. Using reflection to call the
CloneAsync in order to avoid conflicts with the Visual Studio 2017
version. Progress won't be displayed on the status bar, but it appears
prominently on the Team Explorer Home view.