Fix draft appearing after sucessful submit.
In certain situations, the inline comment thread was refreshed before the `DeleteDraft(comment)` line in `PullRequestReviewCommentThreadViewModel.PostComment` was run, meaning that the draft would be displayed _after_ the comment was succesfully submitted.

Work around this by deleting the draft before submitting the comment, and if we get an error re-insert the draft.