Don't change button visibility if repo has changed
There can be delay before IsAGitHubRepo returns its result. If the
repository changes before the call returns, the button can end up with
the visibility for the wrong repository.

This fix only sets the button visibility if the repository hasn't
changed. Button is invalidated multiple times, so one of them returns
the visibility for the correct repository.