Set both affiliations and ownerAffiliations
Explanation by @nicksnyder:

Before this change, the affiliations field was overloaded to mean both
the affiliations that the viewer had with the specified repositories,
as well as the affiliations between the user that the connection was
operating on. This caused a host of issues in various places, and we
opted to separate them out to allow for more flexibility.

In order to fix the above, the proper behavior would be to pass the
full list of arguments to both connections (affiliations and
ownerAffiliations), which should provide you all of the possible
repositories.

https://platform.github.community/t/unable-to-fetch-users-repositories-by-organization-membership/7557/6