Do case sensitive match of owner
When users clone into a directory with the same name as the repository
owner, expect the case of the owner to match.

This avoids a false positive when cloning a repository from the "github"
organization into a direcotry called "GitHub". This is the default directory
that GitHub Desktop uses and could be very anoying.